requirements.txt
Django==1.11.7
psycopg2==2.7.3.2
pytz==2018.3
django-extensions==1.9.9
django-model-utils==3.1.1
djangorestframework==3.7.7
Pillow==5.0.0
celery==4.1.0
